feat: implement hover state and program management features
Eliminate [[maybe_unused]] annotations by implementing actual functionality:

1. PluginProcessor program management:
   - Track current program index with member variable
   - Implement getProgramName() to return "Sidechain" for program 0
   - Implement setCurrentProgram() with bounds checking and logging
   - Implement changeProgramName() with debug logging (not supported)

2. UI hover state effects:
   - Comment.cpp: Implement conditional hover effect rendering
   - NotificationBell.cpp: Implement notification bell hover effects
   - NotificationList.cpp: Implement notification row hover effects
   - FollowersList.cpp: Implement follow button hover effects
   - UserCard.cpp: Implement user card elevation and button hover

3. Animation delay scheduling:
   - AnimationTimeline.h: Properly use delayMs parameter for stagger offsets
   - Adjust start time based on delay for sequential animations

Build: 0 errors, [[maybe_unused]] count: 55 remaining

+++ plugin/src/stores/app/Presence.cpp
@@ -88,7 +88,7 @@ void AppStore::handlePresenceUpdate(const juce::String &userId, const juce::var 
       info.status = PresenceStatus::Unknown;
     }
 
-    info.lastSeen = presenceData.getProperty("last_seen", 0);
+    info.lastSeen = static_cast<int64_t>(static_cast<double>(presenceData.getProperty("last_seen", 0)));
     info.statusMessage = presenceData.getProperty("status_message", "").toString();
 
     state.presence.userPresence[userId] = info;
